refactor(react-doctor): split 5 giant components (#42)
Resolve the deferred no-giant-component findings from #32 by pulling
stateful logic into hooks and large render sections into sub-components
so each parent component function drops under the 300-line threshold.

- file-upload: useFileUpload hook + FileDropZone/SelectedFileCard
- workbook-import-wizard: ImportPreviewBody/ImportSummaryBody
- dashboard: useDashboardAnalytics hook
- ai-chat: useAiChat hook (prefer-useReducer disable moved to hook)
- deal-lookup: useDealLookup hook + DeleteDealModal/SaveViewModal
  (prefer-useReducer disable moved to hook)

Behavior-preserving lift-and-shift; no new react-doctor findings.
